+++ src/components/ValueStreamMap/ValueStreamMap.js
@@ -1,13 +1,9 @@
 import React, { useEffect, useState } from 'react'
-import ReactFlow, {
-  MiniMap,
-  ReactFlowProvider,
-  isNode,
-} from 'react-flow-renderer'
+import ReactFlow, { MiniMap, ReactFlowProvider } from 'react-flow-renderer'
 import { useTheme } from '@material-ui/core/styles'
 import Grid from '@material-ui/core/Grid'
 
-import { getNodeById } from '../../helpers'
+import { getNodeById, isNode } from '../../helpers'
 import { useValueStream } from './valueStreamContext'
 import ConnectionLine from './ConnectionLine'
 import Controls from './Controls'
