+++ src/helpers/index.js
@@ -1,5 +1,4 @@
-export * from './buildNode'
-export * from './buildEdge'
+export * from './elementUtils'
 export * from './utilities'
 export { default as getGraphLayout } from './getGraphLayout'
 export * from './formatVsmData'
